Make a base class
